fix(test): satisfy logic_test_imports + logic_test_file_size on sea_reachable distance tests (Refs #2509)
The earlier `feat(ai): adjacency-distance iteration for
planColonialAcquisition` push added
`packages/colonizethis_logic/test/sea_reachable_provinces_distance_test.dart`
with `package:test/test.dart` directly imported and a 450-physical-line
file. Both tripped CI gates:

- `repo.test_imports` requires `package:colonizethis_test/test.dart`
  first (SPEC/program/test-logging.md) — ct_repo_lint failed at this
  rule before any other file-level checks could run.
- `repo.logic_test_file_size` caps `packages/colonizethis_logic/test/**`
  files at 400 physical lines (SPEC/program/repo-lint.md, "logic test
  file size") — would have failed next.

Minimal compliant fix:

1. Switch the import in
   `sea_reachable_provinces_distance_test.dart` to the canonical
   `package:colonizethis_test/test.dart` shared entrypoint.
2. Split the suite hunkwise into
   `sea_reachable_provinces_distance_test.dart` (4 contracts: direct
   border distance 1, sea-zone path distance 2, canonical NW route
   distance 3, two-paths-shortest-wins) at 241 physical lines and
   `sea_reachable_provinces_distance_part2_test.dart` (3 contracts:
   region filter, BFS non-expansion through foreign provinces,
   determinism) at 254 physical lines. Both files now use the
   `colonizethis_test/test.dart` import and stay well under the 400
   cap.

Verification:

- `dart analyze` of both touched files: no issues.
- `dart test test/sea_reachable_provinces_distance_test.dart
  test/sea_reachable_provinces_distance_part2_test.dart` in
  packages/colonizethis_logic: 7 passed (the same 7 contracts the
  prior single file pinned).
- `dart run tool/ct_repo_lint.dart` with
  `CT_REPO_LINT_INCLUDE_APP=true CT_REPO_LINT_BASE_SHA=$(git merge-base
  HEAD origin/dev) GITHUB_BASE_REF=dev`: all 47 rules pass (previously
  failed at `repo.test_imports`).
- `dart format` clean on both files.

No production changes; SPEC unchanged. Refs #2509.
